Delete blocks of commented code
This shows up in codesearch, and generally makes maintenance more
difficult. This code has been commented out for almost two years
(initial commit), so it seems like it's not really needed.
Especially because the source of this code was updated in the
meanwhile, so it would need updating anyway. Also, it's still
available via git, if necessary.

Change-Id: Ifa3e19fd0ccee20942f1c817ecc4df4b611eaf87