Fix internal usage of WikiText-Parser
This change completly removes `BlueSpice\Social\Parser\WikiText`.
That class was a subclass of the original MediaWiki WikiText-Parser.
It contained very odd changes, like
* leaving out parser stats comments
* omitting images

No actual reason could be found for those. Also subclassing the
MediaWiki WikiText-Parser class is not recommended. Other ways could be
implemented to ensure the desired behavior (e.g. pre- and postprocessing
of the wikitext to parse).

ERM29488

[4.x]

Change-Id: Ibef7e154d5ac029016f506b001f96e94d7737b25